+++ src/Lab5.QMS/DAC/QMSInspectionPlanTest.cs
@@ -3,10 +3,20 @@ using PX.Data;
 
 namespace Lab5.QMS
 {
-    /// <summary>Plan test criteria line. Table UsrQMSInspectionPlanTest (SQL synonym QMSInspectionPlanTest).</summary>
+    /// <summary>
+    /// PX schema-cache table name for <see cref="QMSInspectionPlanTest"/>.
+    /// Synonyms are not tables; PUT Locate/Insert looks up this name.
+    /// </summary>
+    [PXTableName]
+    [Serializable]
+    public class UsrQMSInspectionPlanTest : PXBqlTable, IBqlTable
+    {
+    }
+
+    /// <summary>Plan test criteria line. Table UsrQMSInspectionPlanTest.</summary>
     [Serializable]
     [PXCacheName("Inspection Plan Test")]
-    public class QMSInspectionPlanTest : PXBqlTable, IBqlTable
+    public class QMSInspectionPlanTest : UsrQMSInspectionPlanTest
     {
         #region PlanID
         [PXDBString(30, IsUnicode = true, IsKey = true)]
@@ -20,7 +30,7 @@ namespace Lab5.QMS
 
         #region LineNbr
         [PXDBInt(IsKey = true)]
-        [PXUIField(DisplayName = "Line Nbr", Enabled = false)]
+        [PXUIField(DisplayName = "Line Nbr")]
         public virtual int? LineNbr { get; set; }
         public abstract class lineNbr : PX.Data.BQL.BqlInt.Field<lineNbr> { }
         #endregion